Fix #859, implement a wakeword
This adds a wakeword using Porcupine and its WASM implementation.

It only currently supports the open source wakewords, not any custom wakeword for Firefox Voice.

This defaults to disabled, and must be enabled in the preferences. Users may (must) also select specific wakewords to use.

The actual wakeword detection runs in the recorder page, with a separate parallel media stream (but should naturally pick up the same device as the recorder)

This includes an openPopup() experimental API, which we use to launch listening when the wakeword is received.