Start #728, using a tokenized and formal pattern matching
This splits up words and treats them as tokens to be matched against. It handles stop words and aliases at the matching phase, instead of trying different versions of the string to handle these cases. It also tracks captured and skipped words for ranking matches, instead of having to look at the resulting matches to deduce what happened.

This also adds Jest/Node testing for this module.

The work is incomplete; the code isn't used, is probably missing cases, and I'm not sure how to handle the module situation.