Modularize (#994)
Fix #993, convert most files to Javascript modules

A few modules that are used in content pages, or have other small reasons to not be a normal module, have been left as the old-style modules, globally available without importing. Everything else is a new and proper module, and the browser is handling the module loading natively (without any build step).