Conduct a non-mobile search using `performSearch`
Currently making a search on a non-mobile browser uses the native
extension javascript apis.

Unfortunately the `browser.search.search` method does not return any
relevant info on the search or the newly opened tab. Trying to call
`callScript` to access the search result would throw an error due to
the content script not being loaded.

Next I tried to begin a search by calling `openSearchTab` and passing
the tabId to the native search method. That also failed with a TypeError
since the DOM elements it would search for did not exist.

---
This change will:

[x] conduct a search manually by calling `performSearch`

This now allows us to access the search results and tab info for the
newly created tab.