Allow navigating to the next and previous search results
After conducting an initial search, the search info's index property is
undefined. This caused the extension to believe we were trying to
navigate to a previous result, and fail since there was no previous
result.

This change allows the extension to navigate properly by setting the
initial value of the index prop to 0.