Navigate to next and previous search result (#1180)
Conduct a non-mobile search using `performSearch`

Currently making a search on a non-mobile browser uses the native
extension javascript apis.

Unfortunately the `browser.search.search` method does not return any
relevant info on the search or the newly opened tab. Trying to call
`callScript` to access the search result would throw an error due to
the content script not being loaded.

Next I tried to begin a search by calling `openSearchTab` and passing
the tabId to the native search method. That also failed with a TypeError
since the DOM elements it would search for did not exist.

---
This change will:

[x] conduct a search manually by calling `performSearch`

This now allows us to access the search results and tab info for the
newly created tab.

* Save the search results for a new search

New searches are not being saved into the `tabSearchResults` map.

This change will save the results so that they can be accessed by future
commands

* Allow navigating to the next and previous search results

After conducting an initial search, the search info's index property is
undefined. This caused the extension to believe we were trying to
navigate to a previous result, and fail since there was no previous
result.

This change allows the extension to navigate properly by setting the
initial value of the index prop to 0.

* Navigate to the search page when navigating past the first result

When we are on the first result of a search and ask for "previous
results", the page does not navigate anywhere.

This change will:

[x] navigate to the original search results page
[x] throw an error if the user tries to navigate backwards from the
   search page

* Add a default parameter value to `browserUtil.createTab`

This function call allows creating a new tab with an empty object as the
argument. Yet because it didn't have a default value assigned, it
required always passing an empty object.

This commit assigns an empty object to the function so that it can be
called with no arguments.

* Revert commit 1ab7346e and retrieve search from default search engine

This commit partially reverts 1ab7346e so that it will use the browser's
default search engine to conduct a search. It will then inject the
content script into the new tab so that we can build a `searchInfo`
object that we can store in the `tabSearchResults` map.