Scrape search results from Bing and DuckDuckGo (#1236)
Scrape search results from Bing and DuckDuckGo

Currently the only search provider that we scrape search results from is
google. This means that although search results are displayed for Bing
and DuckDuckGo, the user cannot navigate the results.

This change will:

[x] scrape search results from Bing and DuckDuckGo using a mix of their
BEM classes and html structure.

[x] throws an error if results could not be scraped. This is useful not
only for engines we don't currently scrape, but in case the structure or
classes are changed.

[x] adds a `white-space` css value so that newlines in error messages
are not collapsed.